+++ src/backend/src/main/java/com/becon/opencelium/backend/database/mongodb/repository/MetaDataLogRepository.java
@@ -66,4 +66,6 @@ public interface MetaDataLogRepository extends MongoRepository<LogDataMng, Strin
 
     Optional<LogDataMng> findByExecutionIdAndType(String executionId, String type);
     Optional<LogDataMng> findByFlowIdAndType(String executionId, String type);
+
+    long deleteAllByConnectionId(Long connectionId);
 }